description of compiler warning:

gcc  -std=c99 -ggdb3 -Wall -Werror -Wextra -pedantic  -c main.c -o main.o
main.c: In function ‘main’:
main.c:305:26: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     for (size_t i = 0; i < pass_cnt; ++i) {
                          ^
cc1: all warnings being treated as errors
Makefile:23: recipe for target 'main.o' failed
make: *** [main.o] Error 1

in `strdup`:
changed type of `n` from `int` to `size_t`

possible overflow issue when `strlen(str) + 1 > INT_MAX`

+++ main.c
@@ -62,7 +62,7 @@
  */
 char *strdup(const char *str)
 {
-    int n = strlen(str) + 1;
+    size_t n = strlen(str) + 1;
     char *dup = malloc(n);
     if(dup)
     {
@@ -302,7 +302,7 @@ main(int argc, char **argv)
         exit(EXIT_FAILURE);
     }
 
-    for (size_t i = 0; i < pass_cnt; ++i) {
+    for (long i = 0; i < pass_cnt; ++i) {
         char *pass;
 
         /* We needn't add pass to tree since we free ASAP */
